[Bronze V] Title: 이상한 기호, Time: 208 ms, Memory: 17664 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -3,8 +3,8 @@ import java.util.Scanner;
 class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        int a = scanner.nextInt();
-        int b = scanner.nextInt();
+        long a = scanner.nextLong();
+        long b = scanner.nextLong();
         
         System.out.println((a + b) * (a - b));
     }
